requirements.txt
fastapi[standard]==0.113.0
pydantic==2.10.0
python-dotenv==1.0.1
